Migrate from Bifrost to FinalLogRest interface
