+++ integration-test/src/test/java/org/jboss/pnc/integration/endpoints/BuildPushTest.java
@@ -20,7 +20,7 @@ package org.jboss.pnc.integration.endpoints;
 import org.jboss.arquillian.container.test.api.Deployment;
 import org.jboss.arquillian.container.test.api.RunAsClient;
 import org.jboss.arquillian.junit.Arquillian;
-import org.jboss.pnc.bpm.causeway.BifrostProducer;
+import org.jboss.pnc.bpm.causeway.FinalLogRestProducer;
 import org.jboss.pnc.causewayclient.DefaultCausewayClient;
 import org.jboss.pnc.client.BuildClient;
 import org.jboss.pnc.client.ClientException;
@@ -32,6 +32,7 @@ import org.jboss.pnc.enums.BuildPushStatus;
 import org.jboss.pnc.enums.BuildStatus;
 import org.jboss.pnc.integration.mock.client.BifrostMock;
 import org.jboss.pnc.integration.mock.client.CausewayClientMock;
+import org.jboss.pnc.integration.mock.client.FinalLogRestMock;
 import org.jboss.pnc.integration.setup.Deployments;
 import org.jboss.pnc.integration.setup.RestClientConfiguration;
 import org.jboss.pnc.model.Base32LongID;
@@ -71,8 +72,9 @@ public class BuildPushTest {
         causewayClientJar.addClass(CausewayClientMock.class);
 
         JavaArchive processManager = enterpriseArchive.getAsType(JavaArchive.class, Deployments.BPM_JAR);
-        processManager.deleteClass(BifrostProducer.class);
+        processManager.deleteClass(FinalLogRestProducer.class);
         processManager.addClass(BifrostMock.class);
+        processManager.addClass(FinalLogRestMock.class);
 
         processManager.addAsManifestResource("beans-use-mock-remote-clients.xml", "beans.xml");
